requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
jinja2>=3.1.0
python-multipart>=0.0.9
pillow>=10.0.0
compoundfiles>=0.3
google-genai>=1.0.0
pyembroidery>=1.4.0
